Move auth Callback to the core module
