+++ auth/src/main/java/org/aerogear/mobile/auth/credentials/JwksManager.java
@@ -12,9 +12,9 @@ import android.content.SharedPreferences;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
-import org.aerogear.mobile.auth.Callback;
 import org.aerogear.mobile.auth.configuration.AuthServiceConfiguration;
 import org.aerogear.mobile.auth.configuration.KeycloakConfiguration;
+import org.aerogear.mobile.core.Callback;
 import org.aerogear.mobile.core.MobileCore;
 import org.aerogear.mobile.core.http.HttpRequest;
 import org.aerogear.mobile.core.http.HttpResponse;
@@ -50,9 +50,9 @@ public class JwksManager {
      * cached JWKS found. It will trigger a request to fetch the JWKS in the background if there is
      * no cached key found, or {@link AuthServiceConfiguration#getMinTimeBetweenJwksRequests()} is
      * passed since the key set is requested last time.
-     * 
+     *
      * @param keyCloakConfig the configuration to use to load the JWKS object
-     * 
+     *
      * @return the cached JWKS, or null if it doesn't exist
      */
     public JsonWebKeySet load(final KeycloakConfiguration keyCloakConfig) {
@@ -78,7 +78,7 @@ public class JwksManager {
      * trigger if: 1. forceFetch is set to true, or 2.
      * {@link AuthServiceConfiguration#getMinTimeBetweenJwksRequests()} is passed since the key set
      * is requested last time.
-     * 
+     *
      * @param keycloakConfiguration the configuration of the keycloak server
      * @param forceFetch if set to true, the request will be trigger immediately.
      * @return whether the keys has been fetched or not.
@@ -95,7 +95,7 @@ public class JwksManager {
 
     /**
      * Call the remote endpoint to load the JWKS and save it locally.
-     * 
+     *
      * @param keycloakConfiguration the configuration of the keycloak server
      * @param callback the callback function to be invoked when the request is completed. Can be
      *        null.
@@ -139,7 +139,7 @@ public class JwksManager {
 
     /**
      * Check when the JWKS was requested last time and determine if a request should be sent again.
-     * 
+     *
      * @param keyCloakConfig the configuration of the Keycloak server
      * @return true if the request should be triggered
      */
@@ -159,7 +159,7 @@ public class JwksManager {
 
     /**
      * Save the JWKS content for the given name space locally using SharedPreferences.
-     * 
+     *
      * @param namespace the namespace associated with the JWKS
      * @param jwksContent the content of the JWKS
      */
@@ -177,7 +177,7 @@ public class JwksManager {
 
     /**
      * Build the entry name for the JWKS content
-     * 
+     *
      * @param namespace the namespace associated with the JWKS
      * @return the full entry name
      */
@@ -187,7 +187,7 @@ public class JwksManager {
 
     /**
      * Build the entry name for the last requested date for the JWKS content
-     * 
+     *
      * @param namespace the namespace associated with the JWKS
      * @return the full entry name
      */
