+++ core/src/main/java/org/aerogear/mobile/core/Callback.java
@@ -1,4 +1,4 @@
-package org.aerogear.mobile.auth;
+package org.aerogear.mobile.core;
 
 public interface Callback<T extends Object> {
     default void onSuccess() {}
