+++ example/src/main/java/org/aerogear/mobile/example/ui/AuthDetailsFragment.java
@@ -12,9 +12,9 @@ import android.widget.ListView;
 import android.widget.TextView;
 
 import org.aerogear.mobile.auth.AuthService;
-import org.aerogear.mobile.auth.Callback;
 import org.aerogear.mobile.auth.user.UserPrincipal;
 import org.aerogear.mobile.auth.user.UserRole;
+import org.aerogear.mobile.core.Callback;
 import org.aerogear.mobile.example.R;
 
 import butterknife.BindView;
